river io to usgs and d3d
