+++ mhkit/tests/river/test_io d3d.py
@@ -1,8 +1,6 @@
 from os.path import abspath, dirname, join, isfile, normpath, relpath
 from numpy.testing import assert_array_almost_equal
-from pandas.testing import assert_frame_equal
 import scipy.interpolate as interp
-import matplotlib.pylab as plt
 import mhkit.river as river
 import pandas as pd
 import xarray as xr
@@ -31,45 +29,6 @@ class TestIO(unittest.TestCase):
     @classmethod
     def tearDownClass(self):
         pass
-
-    def test_load_usgs_data_instantaneous(self):
-        file_name = join(datadir, "USGS_08313000_Jan2019_instantaneous.json")
-        data = river.io.usgs.read_usgs_file(file_name)
-
-        self.assertEqual(data.columns, ["Discharge, cubic feet per second"])
-        self.assertEqual(data.shape, (2972, 1))  # 4 data points are missing
-
-    def test_load_usgs_data_daily(self):
-        file_name = join(datadir, "USGS_08313000_Jan2019_daily.json")
-        data = river.io.usgs.read_usgs_file(file_name)
-
-        expected_index = pd.date_range("2019-01-01", "2019-01-31", freq="D")
-        self.assertEqual(data.columns, ["Discharge, cubic feet per second"])
-        self.assertEqual((data.index == expected_index.tz_localize("UTC")).all(), True)
-        self.assertEqual(data.shape, (31, 1))
-
-    def test_request_usgs_data_daily(self):
-        data = river.io.usgs.request_usgs_data(
-            station="15515500",
-            parameter="00060",
-            start_date="2009-08-01",
-            end_date="2009-08-10",
-            data_type="Daily",
-        )
-        self.assertEqual(data.columns, ["Discharge, cubic feet per second"])
-        self.assertEqual(data.shape, (10, 1))
-
-    def test_request_usgs_data_instant(self):
-        data = river.io.usgs.request_usgs_data(
-            station="15515500",
-            parameter="00060",
-            start_date="2009-08-01",
-            end_date="2009-08-10",
-            data_type="Instantaneous",
-        )
-        self.assertEqual(data.columns, ["Discharge, cubic feet per second"])
-        # Every 15 minutes or 4 times per hour
-        self.assertEqual(data.shape, (10 * 24 * 4, 1))
 
     def test_get_all_time(self):
         data = self.d3d_flume_data
